Added indexes route at root
