+++ go.mod
@@ -8,4 +8,6 @@ require (
 	github.com/gofiber/fiber/v2 v2.1.0
 	github.com/pkg/errors v0.9.1
 	github.com/thedevsaddam/gojsonq v2.3.0+incompatible // indirect
+	github.com/tidwall/gjson v1.6.8 // indirect
+	github.com/tidwall/pretty v1.0.5 // indirect
 )
