+++ internal/handlers/indexes.go
@@ -1,0 +1,1 @@
+package handlers
