+++ lambda/armies/main.go
@@ -6,8 +6,8 @@ import (
 	"github.com/aws/aws-lambda-go/events"
 	"github.com/aws/aws-lambda-go/lambda"
 	fiberadapter "github.com/awslabs/aws-lambda-go-api-proxy/fiber"
-	"github.com/brittonhayes/warhammer-aos/armies"
 	"github.com/brittonhayes/warhammer-aos/internal/handlers"
+	"github.com/brittonhayes/warhammer-aos/pkg/armies"
 )
 
 var fiberLambda *fiberadapter.FiberLambda
