+++ pkg/armies/README.md
@@ -1,0 +1,162 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# armies
+
+```go
+import "github.com/brittonhayes/warhammer-aos/pkg/armies"
+```
+
+package armies contains the types\, methods\, and interfaces for interacting with Warhammer Age of Sigmar Armies
+
+## Index
+
+- [func Handler() func(ctx *fiber.Ctx) error](<#func-handler>)
+- [type Ability](<#type-ability>)
+- [type Armies](<#type-armies>)
+  - [func (a *Armies) Reply() *handlers.Response](<#func-armies-reply>)
+- [type Army](<#type-army>)
+  - [func (a *Army) Reply() *handlers.Response](<#func-army-reply>)
+  - [func (a *Army) Unit(name string) string](<#func-army-unit>)
+- [type MeleeWeapon](<#type-meleeweapon>)
+- [type MissileWeapon](<#type-missileweapon>)
+- [type Service](<#type-service>)
+- [type Unit](<#type-unit>)
+
+
+## func Handler
+
+```go
+func Handler() func(ctx *fiber.Ctx) error
+```
+
+Handler returns an http response all of the armies as a JSON
+
+## type Ability
+
+Ability is a skill possessed by the parent unit
+
+```go
+type Ability struct {
+    Name string `json:"name"`
+    Desc string `json:"desc"`
+}
+```
+
+## type Armies
+
+Armies is a helper type to allow receiver methods for a slice of Army
+
+```go
+type Armies []Army
+```
+
+<details><summary>Example</summary>
+<p>
+
+Create a new handler and add a function to the route
+
+```go
+package main
+
+import (
+	"github.com/brittonhayes/warhammer-aos/internal/handlers"
+	"github.com/brittonhayes/warhammer-aos/pkg/armies"
+)
+
+func main() {
+	_ = handlers.New().AddRoute("*", armies.Handler()).Build()
+}
+```
+
+</p>
+</details>
+
+### func \(\*Armies\) Reply
+
+```go
+func (a *Armies) Reply() *handlers.Response
+```
+
+Reply returns a list of armies formatted as an API response
+
+## type Army
+
+Army is the parent structure that all nested units belong to
+
+```go
+type Army struct {
+    Army  string `json:"army"`
+    Units []Unit `json:"units"`
+}
+```
+
+### func \(\*Army\) Reply
+
+```go
+func (a *Army) Reply() *handlers.Response
+```
+
+Reply returns an army formatted as an API response
+
+### func \(\*Army\) Unit
+
+```go
+func (a *Army) Unit(name string) string
+```
+
+## type MeleeWeapon
+
+MeleeWeapon is a weapon that is used for close quarters combat
+
+```go
+type MeleeWeapon struct {
+    Name    string `json:"name"`
+    Range   string `json:"range"`
+    Attacks string `json:"attacks"`
+    ToHit   string `json:"to_hit"`
+    ToWound string `json:"to_wound"`
+    Rend    string `json:"rend"`
+    Damage  string `json:"damage"`
+}
+```
+
+## type MissileWeapon
+
+MissileWeapon is a weapon that is used for ranged attacks
+
+```go
+type MissileWeapon struct {
+    Name    string `json:"name"`
+    Range   string `json:"range"`
+    Attacks string `json:"attacks"`
+    ToHit   string `json:"to_hit"`
+    ToWound string `json:"to_wound"`
+    Rend    string `json:"rend"`
+    Damage  string `json:"damage"`
+}
+```
+
+## type Service
+
+Service covers all available methods of the Armies package
+
+```go
+type Service interface {
+    Reply() *handlers.Response
+}
+```
+
+## type Unit
+
+Unit is an individual unit in Warhammer
+
+```go
+type Unit struct {
+    Name string `json:"name"`
+    Size string `json:"size"`
+}
+```
+
+
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
