+++ pkg/indexes/README.md
@@ -1,0 +1,56 @@
+<!-- Code generated by gomarkdoc. DO NOT EDIT -->
+
+# indexes
+
+```go
+import "github.com/brittonhayes/warhammer-aos/pkg/indexes"
+```
+
+## Index
+
+- [func Handler() func(ctx *fiber.Ctx) error](<#func-handler>)
+- [type Index](<#type-index>)
+  - [func (i Index) List() []string](<#func-index-list>)
+  - [func (i *Index) Reply() *handlers.Response](<#func-index-reply>)
+
+
+## func Handler
+
+```go
+func Handler() func(ctx *fiber.Ctx) error
+```
+
+Handler returns an http response all of the indexes as a JSON
+
+## type Index
+
+```go
+type Index string
+```
+
+```go
+const (
+    RootIndex   Index = "/api"
+    ArmiesIndex       = "/api/armies"
+)
+```
+
+### func \(Index\) List
+
+```go
+func (i Index) List() []string
+```
+
+List lists out the the available API routes
+
+### func \(\*Index\) Reply
+
+```go
+func (i *Index) Reply() *handlers.Response
+```
+
+Reply returns the indexes formatted as an API response
+
+
+
+Generated by [gomarkdoc](<https://github.com/princjef/gomarkdoc>)
