+++ serverless.yml
@@ -30,10 +30,15 @@ custom:
     securityPolicy: tls_1_2
 
 functions:
+  indexes:
+    handler: ./bin/indexes
+    events:
+      - http:
+          path: /
+          method: GET
   armies:
     handler: ./bin/armies
     events:
       - http:
           path: armies
           method: GET
-
